+++ CHANGELOG.md
@@ -6,9 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added optional `name` parameter to React `DataProvider` which will use the
+  supplied string as the state-name.
+- Added optional `name` parameter to `Storage` and `StateInstance` `getNested`,
+  `getNestedOrCreate`, and `removeNested`.
+
 ### Changes
 - `Subscription`, `SubscriptionMap` and `MessageFilter` are now generic over
   `<M: Message>` instead of just using `Message`.
+- `Snapshot` `defName` is renamed to `id`.
 
 ## [0.0.6] - 2019-06-07
 ### Changed